requirements.txt
Flask
python-dotenv
shortuuid
gunicorn
